+++ src/tvboptim/experimental/network_dynamics/core/__init__.py
@@ -1,14 +1,14 @@
 """Core components of  Network Dynamics."""
 
 from .bunch import Bunch
-from .heterogeneous import DynamicsGroup, HeterogeneousNetwork, SignalRoute
+from .heterogeneous import HeterogeneousNetwork, NodeGroup, SignalRoute
 from .network import Network
 from .observation import GroupObservation
 from .readout import Readout
 
 __all__ = [
     "Bunch",
-    "DynamicsGroup",
+    "NodeGroup",
     "HeterogeneousNetwork",
     "GroupObservation",
     "Network",